Move module analysis.utils to top level
